+++ tests/unit/domestic/factories.py
@@ -7,6 +7,7 @@ class DomesticHomePageFactory(wagtail_factories.PageFactory):
 
     title = 'homepage'
     live = True
+    slug = 'homepage'
 
     class Meta:
         model = models.DomesticHomePage